One duckdb string type to rule them all
