+++ vortex-duckdb/src/duckdb/mod.rs
@@ -8,6 +8,7 @@ mod copy_function;
 mod data;
 mod data_chunk;
 mod database;
+mod ddb_string;
 mod expr;
 pub mod footer_cache;
 mod logical_type;
@@ -16,7 +17,6 @@ mod object_cache;
 mod query_result;
 mod scalar_function;
 mod selection_vector;
-mod string;
 mod table_filter;
 mod table_function;
 mod value;
@@ -39,7 +39,6 @@ pub use object_cache::*;
 pub use query_result::*;
 pub use scalar_function::*;
 pub use selection_vector::*;
-pub use string::*;
 pub use table_filter::*;
 pub use table_function::*;
 pub use value::*;
